requirements.txt
firebase-admin==5.2.0
google-auth-oauthlib==0.4.6
Pillow==9.0.1
pandas==1.4.1
streamlit-authenticator==0.2.1
geonamescache==1.5.0